CHORE: UFCS for orbs
