+++ src/granad/_sim.py
@@ -1,4 +1,3 @@
-from flax import struct
 from dataclasses import dataclass, fields, field
 import numpy as np
 import jax.numpy as jnp
@@ -7,56 +6,19 @@ from functools import wraps
 import jax
 import itertools
 import warnings
-from granad import *
 import json
-from _shapes import shapes
 import uuid
+
+# TODO: hmmm
+from ._core import *
+from . import _observables 
 
 # TODO: deprecated but idk tbh
 from pkg_resources import resource_filename, resource_stream
 
 # TODO JAXIFX as much as possible
-@struct.dataclass
-class Stack:
-    """A stack of orbitals.
 
-    - `hamiltonian`: TB-Hamiltonian of the system.
-    - `coulomb`: Coulomb matrix of the system.
-    - `rho_0`: Density matrix corresponding to the initial state. If a simulation is interrupted at time t, then it can be resumed using the density matrix at time t as rho_0.
-    - `rho_stat`: Density matrix corresponding to the final state.
-    - `energies`: Independent particle energies as given by the eigenvalues of the TB-hamiltonian.
-    - `eigenvectors`: Eigenvectors of the TB-Hamiltonian. These are used to transform between *site* and *energy* basis.
-    - `positions`: Positions of the orbitals, $N \\times 3$ - dimensional.
-    - `unique_ids`: Collection of unique strings. Each string identifies an orbital type occuring in the stack. Typically something like `"pz_graphene"`.
-    - `ids`: The integer at the n-th position corresponds to the type of the n-th orbital via its index in `unique_ids`.
-    - `eps`: numerical precision threshold for identifiying degeneracies in the eigenenergies.
-    - `homo`: index of the homo, such that stack.energies[homo] is the energy of the ground state homo.
-    - `electrons`: total number of electrons in the system.
-    - `from_state`: the state from which electrons have been taken for the initial excited transition, the state translates to HOMO - from_state
-    - `to_state`: the state into which electrons haven been put for the initial excited transition, the state translates to HOMO + from_state
-    - `beta`: corresponds to $(k_B T)^{-1}$
-    """
-
-    hamiltonian: jax.Array
-    coulomb: jax.Array
-    rho_0: jax.Array
-    rho_stat: jax.Array
-    energies: jax.Array
-    eigenvectors: jax.Array
-    positions: jax.Array
-    unique_ids: list[str] # really needed? only once for dipoles
-    ids: jax.Array  # really needed? only once for dipoles
-    eps: float
-    homo: int
-    electrons: int
-    from_state: int
-    to_state: int
-    excited_electrons: int
-    beta: float
-    spin_degeneracy: float
-    transitions: dict
-
-# TODO: this is not thread-safe
+# TODO: this is not thread-safe... but do I care?
 class Counter:
     _counter = 0
 
@@ -331,13 +293,6 @@ class OrbitalList:
 
         # list of fields that are directly looked up from the stack
         self._stack_fields = [ n.name for n in fields(Stack)]
-
-        # TODO: ufff
-        self._observable_functions = self._get_class_method_names()
-
-    @staticmethod
-    def _get_class_method_names( cls ):
-        return [func.__name__ for func in dir(cls) if callable(getattr(cls, func)) and not func.startswith('__')]
         
     def __len__(self):
         return len(self._list)
@@ -389,11 +344,8 @@ class OrbitalList:
         del self._list[position]
 
     def __getattr__(self, item):
-        if not item in (self._stack_fields + self._functions):        
+        if not item in self._stack_fields:        
             raise AttributeError(f"'{type(self).__name__}' object has no attribute '{item}'")
-        
-        if item in self._plotting_functions:
-            return getattr( self.dummy, item )
         
         if self._recompute_stack:
             self.build()
@@ -550,3 +502,8 @@ class OrbitalList:
 
     def _hamiltonian_coulomb_slater_koster():
         return NotImplemented
+
+for name in dir(_observables):
+    method = getattr(_observables, name)
+    if callable(method):
+        setattr(OrbitalList, name, method)
